requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision
opencv-python>=4.8.0
albumentations>=1.3.0
scikit-image>=0.21.0

# 3D processing
trimesh>=3.23.0
open3d>=0.17.0
scipy>=1.11.0

# Data processing
numpy>=1.24.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
omegaconf>=2.3.0

# Logging (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0